+++ openfisca_core/holders.py
@@ -2,6 +2,8 @@
 
 
 from __future__ import division
+import warnings
+import os
 
 import numpy as np
 
@@ -55,12 +57,22 @@ class Holder(object):
     formula = None
     formula_output_period_by_requested_period = None
 
-    def __init__(self, simulation, column = None):
+    def __init__(self, simulation = None, column = None, entity = None):
         assert column is not None
         assert self.column is None
+        if simulation is not None:
+            warnings.warn(
+                u"The Holder(simulation, column) constructor has been deprecated. "
+                u"Please use Holder(entity = entity, column = column) instead.",
+                Warning
+                )
+            self.simulation = simulation
+            self.entity = simulation.get_entity(column.entity)
+        else:
+            self.entity = entity
+            self.simulation = entity.simulation
         self.column = column
-        self.simulation = simulation
-        self.entity = self.simulation.entities[self.column.entity.key]
+        self.buffer = {}
 
     @property
     def array(self):
@@ -245,6 +257,21 @@ class Holder(object):
         return formula.real_formula
 
     def set_input(self, period, array):
+        if period.unit == ETERNITY and self.column.definition_period != ETERNITY:
+            error_message = os.linesep.join([
+                u'Unable to set a value for variable {0} for ETERNITY.',
+                u'{0} is only defined for {1}s. Please adapt your input.',
+                ]).format(
+                    self.column.name,
+                    self.column.definition_period
+                ).encode('utf-8')
+            raise PeriodMismatchError(
+                self.column.name,
+                period,
+                self.column.definition_period,
+                error_message
+                )
+
         self.formula.set_input(period, array)
 
     def put_in_cache(self, value, period, extra_params = None):
@@ -255,12 +282,23 @@ class Holder(object):
                 raise ValueError('A period must be specified to put values in cache, except for variables with ETERNITY as as period_definition.')
             if ((self.column.definition_period == MONTH and period.unit != periods.MONTH) or
                (self.column.definition_period == YEAR and period.unit != periods.YEAR)):
-                raise ValueError(u'Unable to set a value for variable {0} for {1}-long period {2}. {0} is only defined for {3}s. Please adapt your input. If you are the maintainer of {0}, you can consider adding it a set_input attribute to enable automatic period casting.'.format(
+                error_message = os.linesep.join([
+                    u'Unable to set a value for variable {0} for {1}-long period {2}.',
+                    u'{0} is only defined for {3}s. Please adapt your input.',
+                    u'If you are the maintainer of {0}, you can consider adding it a set_input attribute to enable automatic period casting.'
+                    ]).format(
+                        self.column.name,
+                        period.unit,
+                        period,
+                        self.column.definition_period
+                    ).encode('utf-8')
+
+                raise PeriodMismatchError(
                     self.column.name,
-                    period.unit,
                     period,
-                    self.column.definition_period
-                    ).encode('utf-8'))
+                    self.column.definition_period,
+                    error_message
+                    )
 
         if (simulation.opt_out_cache and
                 simulation.tax_benefit_system.cache_blacklist and
@@ -343,3 +381,11 @@ class Holder(object):
         array = np.empty(array_size, dtype = self.column.dtype)
         array.fill(self.column.default)
         return array
+
+
+class PeriodMismatchError(ValueError):
+    def __init__(self, variable_name, period, definition_period, message):
+        self.variable_name = variable_name
+        self.period = period
+        self.definition_period = definition_period
+        ValueError.__init__(self, message)