+++ openfisca_web_api_preview/loader/__init__.py
@@ -1,14 +1,9 @@
 # -*- coding: utf-8 -*-
 
-import os
-
-import yaml
-
-from .tax_benefit_system import build_tax_benefit_system
-from .parameters import build_parameters
-from .variables import build_variables
-
-OPEN_API_CONFIG_FILE = os.path.join(os.path.dirname(os.path.abspath(__file__)), os.path.pardir, 'openAPI.yml')
+from tax_benefit_system import build_tax_benefit_system
+from parameters import build_parameters
+from variables import build_variables
+from spec import build_openAPI_specification
 
 
 def extract_description(items):
@@ -18,22 +13,12 @@ def extract_description(items):
         }
 
 
-def build_openAPI_specification(country_package_metadata):
-    file = open(OPEN_API_CONFIG_FILE, 'r')
-    spec = yaml.load(file)
-    country_package_name = country_package_metadata['name'].title()
-    spec['info']['title'] = spec['info']['title'].replace("{COUNTRY_PACKAGE_NAME}", country_package_name)
-    spec['info']['description'] = spec['info']['description'].replace("{COUNTRY_PACKAGE_NAME}", country_package_name)
-    spec['host'] = os.environ.get('SERVER_NAME')
-    return spec
-
-
 def build_data(country_package_name):
     tax_benefit_system = build_tax_benefit_system(country_package_name)
     country_package_metadata = tax_benefit_system.get_package_metadata()
     parameters = build_parameters(tax_benefit_system)
     variables = build_variables(tax_benefit_system, country_package_metadata)
-    openAPI_spec = build_openAPI_specification(country_package_metadata)
+    openAPI_spec = build_openAPI_specification(tax_benefit_system, country_package_metadata)
     return {
         'tax_benefit_system': tax_benefit_system,
         'country_package_metadata': tax_benefit_system.get_package_metadata(),