+++ tests/core/test_extra_params.py
@@ -59,7 +59,7 @@ simulation = tax_benefit_system.new_scenario().init_from_attributes(
     ).new_simulation(debug = True)
 formula_1_result = simulation.calculate('formula_1', period = reference_period)
 formula_2_result = simulation.calculate('formula_2', period = reference_period)
-formula_3_holder = simulation.holder_by_name['formula_3']
+formula_3_holder = simulation.person.get_holder('formula_3')
 
 
 def test_cache():